remove BaseSideMenuPage.java, RF code
