+++ src/test/java/school/redrover/model/HomePage.java
@@ -7,14 +7,14 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import school.redrover.model.base.BaseSideMenuPage;
+import school.redrover.model.base.BasePage;
 import school.redrover.runner.TestUtils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class HomePage extends BaseSideMenuPage<HomePage> {
+public class HomePage extends BasePage {
 
     @FindBy(linkText = "Create a job")
     private WebElement createAJobLink;
@@ -472,12 +472,6 @@ public class HomePage extends BaseSideMenuPage<HomePage> {
         hoverOverElement(tooltip);
 
         return tooltip.isDisplayed();
-    }
-
-    public UserConfigurePage openUserConfigurations() {
-        return clickPeople()
-                .clickUserIdLink()
-                .clickConfigureOnSidebar();
     }
 
     public List<String> getSidebarMenuList() {
