+++ src/test/java/school/redrover/model/UserPage.java
@@ -3,12 +3,16 @@ package school.redrover.model;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import school.redrover.model.base.BaseSideMenuPage;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import school.redrover.model.base.BasePage;
 
-public class UserPage extends BaseSideMenuPage<UserPage> {
+public class UserPage extends BasePage {
 
     @FindBy(css = "#description + div")
     private WebElement jenkinsUserID;
+
+    @FindBy(linkText = "Configure")
+    private WebElement configureOnSidebar;
 
     public UserPage(WebDriver driver) { super(driver); }
 
@@ -16,5 +20,12 @@ public class UserPage extends BaseSideMenuPage<UserPage> {
 
         return jenkinsUserID.getText().replace("Jenkins User ID:", "").trim();
     }
+
+    public UserConfigurePage clickConfigureOnSidebar() {
+        getWait5().until(ExpectedConditions.visibilityOf(configureOnSidebar)).click();
+
+        return new UserConfigurePage(getDriver());
+    }
+
 
 }
